Fix issues with the signedness of char
Many apis in apr use the C type char, depending on the architecture
this may map to either the i8 (x86-64, loong64) or the u8 (arm64,
ppc64el, s390x, ppc64) type in rust. The standard library offers
a type alias c_char which is an alias for either i8 or u8 depending
on the architecture.

However, rust-apr has a bunch of hardcoded use of i8, which causes
it to fail to build on architectures where char is unsigned.